Fix for issue 2 "mixed-case db_column= names in PostgreSQL cause invalid queries"
Basically we make sure that the id column we attach to let us link the related fields back in has a mangled name so it doesn't stop the regular id column used in filter() from not being quoted (which causes the postgres issue).

This also merges in the use of quote_name from akaihola.