Fixed failing test in Django 1.3 with non-integer pk
Caused due to apparent regression in Django 1.3, that means that in queries with entities with non-integer keys don't seem to work anymore, but only when we pass in a list of either the instances/ids and not a queryset.

Handling this by explicitly making in query reference the pk of the relevant entity.