Detecting when we've been passed full reverse fk name and handling
get_field_by_name (by default) expects field name for reverse foreign key managers to not include the _set suffix, so in batch_select we would have to do:

    Section.objects.batch_select('entry')

Instead of:

    Section.objects.batch_select('entry_set')

Even though entry_set is the real field name - this appears to be slightly odd behaviour in Django, as if we specify relate_name in the fk everything works as expected.  To deal with this I now check whether the field name passed in ends with a _set and remove the suffix if I don't otherwise find the field under the original name.

This should make using batch select feel a bit more predictable - less surprising.